+++ evealert/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "4.1.0"
+__version__ = "4.2.0"
 __title__ = "EVE Alert"